Fixed a few tests
